reorganize everything out of a single file
