+++ OpenDreamServer/Dream/Objects/DreamObjectDefinition.cs
@@ -9,9 +9,9 @@ namespace OpenDreamServer.Dream.Objects {
     class DreamObjectDefinition {
         public DreamPath Type;
         public IDreamMetaObject MetaObject = null;
-        public DreamProc InitializionProc = null;
-        public Dictionary<string, DreamProc> Procs { get; private set; } = new();
-        public Dictionary<string, DreamProc> OverridingProcs { get; private set; } = new();
+        public Proc InitializionProc = null;
+        public Dictionary<string, Proc> Procs { get; private set; } = new();
+        public Dictionary<string, Proc> OverridingProcs { get; private set; } = new();
         public Dictionary<string, DreamValue> Variables { get; private set; } = new();
         public Dictionary<string, DreamGlobalVariable> GlobalVariables { get; private set; } = new();
 
@@ -29,11 +29,11 @@ namespace OpenDreamServer.Dream.Objects {
 
             CopyVariablesFrom(copyFrom);
 
-            foreach (KeyValuePair<string, DreamProc> proc in copyFrom.Procs) {
+            foreach (KeyValuePair<string, Proc> proc in copyFrom.Procs) {
                 Procs.Add(proc.Key, proc.Value);
             }
 
-            foreach (KeyValuePair<string, DreamProc> proc in copyFrom.OverridingProcs) {
+            foreach (KeyValuePair<string, Proc> proc in copyFrom.OverridingProcs) {
                 OverridingProcs.Add(proc.Key, proc.Value);
             }
         }
@@ -50,7 +50,7 @@ namespace OpenDreamServer.Dream.Objects {
             Variables[variableName] = value;
         }
 
-        public void SetProcDefinition(string procName, DreamProc proc) {
+        public void SetProcDefinition(string procName, Proc proc) {
             if (HasProc(procName)) {
                 proc.SuperProc = GetProc(procName);
                 OverridingProcs[procName] = proc;
@@ -59,7 +59,7 @@ namespace OpenDreamServer.Dream.Objects {
             }
         }
 
-        public void SetNativeProc(NativeProcHandler func) {
+        public void SetNativeProc(NativeProc.NativeProcHandler func) {
             List<Attribute> attributes = new(func.GetInvocationList()[0].Method.GetCustomAttributes());
             DreamProcAttribute procAttribute = (DreamProcAttribute)attributes.Find(attribute => attribute is DreamProcAttribute);
             if (procAttribute == null) throw new ArgumentException();
@@ -69,15 +69,15 @@ namespace OpenDreamServer.Dream.Objects {
             SetProcDefinition(procAttribute.Name, proc);
         }
 
-        public DreamProc GetProc(string procName) {
-            if (TryGetProc(procName, out DreamProc proc)) {
+        public Proc GetProc(string procName) {
+            if (TryGetProc(procName, out Proc proc)) {
                 return proc;
             } else {
                 throw new Exception("Object type '" + Type + "' does not have a proc named '" + procName + "'");
             }
         }
 
-        public bool TryGetProc(string procName, out DreamProc proc) {
+        public bool TryGetProc(string procName, out Proc proc) {
             if (OverridingProcs.TryGetValue(procName, out proc)) {
                 return true;
             } else if (Procs.TryGetValue(procName, out proc)) {
