+++ OpenDreamServer/Dream/Procs/DMProc.cs
@@ -5,60 +5,10 @@ using OpenDreamServer.Dream.Objects;
 using OpenDreamShared.Dream.Procs;
 
 namespace OpenDreamServer.Dream.Procs {
-    enum ProcStatus {
-        Returned,
-        Deferred,
-        Called,
-    }
-
-    delegate DreamValue NativeProcHandler(DreamObject src, DreamObject usr, DreamProcArguments arguments);
-
-    abstract class DreamProc {
-        public string Name { get; }
-
-        // This is currently publically settable because the loading code doesn't know what our super is until after we are instantiated
-        public DreamProc SuperProc { set; get; }
-
-        public List<String> ArgumentNames { get; }
-        public List<DMValueType> ArgumentTypes { get; }
-
-        protected DreamProc(string name, DreamProc superProc, List<String> argumentNames, List<DMValueType> argumentTypes) {
-            Name = name;
-            SuperProc = superProc;
-            ArgumentNames = argumentNames ?? new();
-            ArgumentTypes = argumentTypes ?? new();
-        }
-
-        public abstract ProcState CreateState(ExecutionContext context, DreamObject src, DreamObject usr, DreamProcArguments arguments);
-
-        // Execute this proc. This will behave as if the proc has `set waitfor = 0`
-        public DreamValue Run(DreamObject src, DreamProcArguments arguments, DreamObject usr = null) {
-            var context = new ExecutionContext();
-            var state = CreateState(context, src, usr, arguments);
-            context.PushProcState(state);
-            return context.Resume();
-        }
-    }
-
-    class NativeProc : DreamProc {
-        public NativeProcHandler Func { get; }
-
-        public NativeProc(string name, DreamProc superProc, List<String> argumentNames, List<DMValueType> argumentTypes, NativeProcHandler func)
-            : base(name, superProc, argumentNames, argumentTypes)
-        {
-            Func = func;
-        }
-
-        public override TrivialNativeProcState CreateState(ExecutionContext context, DreamObject src, DreamObject usr, DreamProcArguments arguments)
-        {
-            return new TrivialNativeProcState(this, context, src, usr, arguments);
-        }
-    }
-
-    class DMProc : DreamProc {
+    class DMProc : Proc {
         public byte[] Bytecode { get; }
 
-        public DMProc(string name, DreamProc superProc, List<String> argumentNames, List<DMValueType> argumentTypes, byte[] bytecode)
+        public DMProc(string name, Proc superProc, List<String> argumentNames, List<DMValueType> argumentTypes, byte[] bytecode)
             : base(name, superProc, argumentNames, argumentTypes)
         {
             Bytecode = bytecode;
@@ -67,101 +17,6 @@ namespace OpenDreamServer.Dream.Procs {
         public override DMProcState CreateState(ExecutionContext context, DreamObject src, DreamObject usr, DreamProcArguments arguments)
         {
             return new DMProcState(this, context, src, usr, arguments);
-        }
-    }
-
-    abstract class ProcState {
-        public ExecutionContext Context { get; }
-        public DreamValue Result { set; get; } = DreamValue.Null;
-        
-        public ProcState(ExecutionContext context) {
-            Context = context;
-        }
-
-        public abstract DreamProc Proc { get; }
-        public abstract ProcStatus Resume();
-
-        // Most implementations won't require this, so give it a default
-        public virtual void ReturnedInto(DreamValue value) {}
-    }
-
-    // Used for synchronous procs that won't need to persist any state between executions (as they only execute once.)
-    class TrivialNativeProcState : ProcState
-    {
-        public DreamObject Src;
-        public DreamObject Usr;
-        public DreamProcArguments Arguments;
-        
-        private NativeProc _proc;
-        public override DreamProc Proc => _proc;
-
-        public TrivialNativeProcState(NativeProc proc, ExecutionContext context, DreamObject src, DreamObject usr, DreamProcArguments arguments)
-            : base(context)
-        {
-            _proc = proc;
-            Src = src;
-            Usr = usr;
-            Arguments = arguments;
-        }
-
-        public override ProcStatus Resume()
-        {
-            Result = _proc.Func.Invoke(Src, Usr, Arguments);
-            return ProcStatus.Returned;
-        }
-
-        public override void ReturnedInto(DreamValue value) {
-            
-        }
-    }
-
-    class ExecutionContext {
-        private ProcState _current; 
-        private Stack<ProcState> _stack = new();
-
-        public DreamValue Resume() {
-            if (!Program.IsMainThread) {
-                throw new InvalidOperationException();
-            }
-
-            while (_current != null) {
-                // _current.Resume may mutate our state!!!
-                switch (_current.Resume()) {
-                    // Our top-most proc just returned a value
-                    case ProcStatus.Returned:
-                        var returned = _current.Result;
-
-                        // If our stack is empty, the context has finished execution
-                        // so we can return the result to our native caller
-                        if (!_stack.TryPop(out _current)) {
-                            return returned;
-                        }
-
-                        // ... otherwise we just push the return value onto the dm caller's stack
-                        _current.ReturnedInto(returned);
-                        break;
-
-                    // The context is done executing for now
-                    case ProcStatus.Deferred:
-                        // We return the current return value here even though it may not be the final result
-                        return _current.Result;
-
-                    // Our top-most proc just called a function
-                    // This means _current has changed!
-                    case ProcStatus.Called:
-                        // Nothing to do. The loop will call into _current.Resume for us.
-                        break;
-                }
-            }
-
-            throw new InvalidOperationException();
-        }
-
-        public void PushProcState(ProcState state) {
-            if (_current != null) {
-                _stack.Push(_current);
-            }
-            _current = state;
         }
     }
 
@@ -262,7 +117,7 @@ namespace OpenDreamServer.Dream.Procs {
         private int _pc = 0;
 
         private DMProc _proc;
-        public override DreamProc Proc => _proc;
+        public override Proc Proc => _proc;
 
         public DMProcState(DMProc proc, ExecutionContext context, DreamObject instance, DreamObject usr, DreamProcArguments arguments)
             : base(context)
@@ -337,7 +192,7 @@ namespace OpenDreamServer.Dream.Procs {
             Result = value;
         }
 
-        public void Call(DreamProc proc, DreamObject src, DreamProcArguments arguments) {
+        public void Call(Proc proc, DreamObject src, DreamProcArguments arguments) {
             var state = proc.CreateState(Context, src, Usr, arguments);
             Context.PushProcState(state);
         }
@@ -413,5 +268,4 @@ namespace OpenDreamServer.Dream.Procs {
         }
         #endregion
     }
-
 }
