+++ OpenDreamServer/Dream/Procs/Native/DreamProcNativeRoot.cs
@@ -1247,7 +1247,7 @@ namespace OpenDreamServer.Dream.Procs.Native {
                     DreamPath objectTypePath = typePath.AddToPath("..");
                     DreamObjectDefinition objectDefinition = Program.DreamObjectTree.GetObjectDefinitionFromPath(objectTypePath);
 
-                    foreach (KeyValuePair<string, DreamProc> proc in objectDefinition.Procs) {
+                    foreach (KeyValuePair<string, Proc> proc in objectDefinition.Procs) {
                         list.AddValue(new DreamValue(proc.Key));
                     }
                 } else {
